initial implementation
crypted is heavily inspired by the old droidian encryption service. droidian encryption service did its job but it was very bare bones and the code tries really hard to be glib compliant and pretty, but that makes it extra ugly and unreadable. we are also switching to furios VG which required me to redo parts of the old encryption service to work with multiple volume group and partition names while keeping the configuration code. old devices remaining on the old VG (pre 13.1.0) but they will remain compatible but marked as legacy

the ramdisk helper also built against glib which made ramdisk bigger than it should've been. in general, it would be more work to redo that service than to rewrite it entirely

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>